refactor: ♻️ Move Hooks into their own module
